requirements.txt
boto3
chromadb
nltk
pyyaml
python-dotenv
slack_bolt